+++ YoloOnnxWinform/YoloDotNet/Modules/12/PoseEstimationModuleV12.cs
@@ -23,7 +23,7 @@ namespace YoloDotNet.Modules.V12
             _poseEstimationModuleV8 = new PoseEstimationModuleV8(_yoloCore);
         }
 
-        public List<PoseEstimation> ProcessImage<T>(T image, double confidence, double pixelConfidence, double iou)
+        public List<PoseEstimation> ProcessImage<T>(T image, float confidence, float pixelConfidence, float iou)
             => _poseEstimationModuleV8.ProcessImage(image, confidence, pixelConfidence, iou);
 
         #region Helper methods
