+++ YoloOnnxWinform/YoloDotNet/Modules/Interfaces/IClassificationModule.cs
@@ -6,6 +6,6 @@ namespace YoloDotNet.Modules.Interfaces
 {
     public interface IClassificationModule : IModule
     {
-        List<Classification> ProcessImage<T>(T image, double classes, double pixelConfidence, double iou);
+        List<Classification> ProcessImage<T>(T image, float classes, float pixelConfidence, float iou);
     }
 }
