+++ YoloOnnxWinform/YoloDotNet/Modules/Interfaces/IOBBDetectionModule.cs
@@ -6,6 +6,6 @@ namespace YoloDotNet.Modules.Interfaces
 {
     internal interface IOBBDetectionModule : IModule
     {
-        List<OBBDetection> ProcessImage<T>(T image, double confidence, double pixelConfidence,double iou);
+        List<OBBDetection> ProcessImage<T>(T image, float confidence, float pixelConfidence, float iou);
     }
 }
