+++ YoloOnnxWinform/YoloDotNet/Modules/Interfaces/IPoseEstimationModule.cs
@@ -6,6 +6,6 @@ namespace YoloDotNet.Modules.Interfaces
 {
     internal interface IPoseEstimationModule : IModule
     {
-        List<PoseEstimation> ProcessImage<T>(T image, double confidence, double pixelConfidence, double iou);
+        List<PoseEstimation> ProcessImage<T>(T image, float confidence, float pixelConfidence, float iou);
     }
 }
