+++ YoloOnnxWinform/YoloDotNet/Modules/V8E/SegmentationModuleV8E.cs
@@ -19,7 +19,7 @@ namespace YoloDotNet.Modules.V8E
             _segmentationModuleV8 = new SegmentationModuleV8(_yoloCore);
         }
 
-        public List<Segmentation> ProcessImage<T>(T image, double confidence, double pixelConfidence, double iou)
+        public List<Segmentation> ProcessImage<T>(T image, float confidence, float pixelConfidence, float iou)
             => _segmentationModuleV8.ProcessImage(image, confidence, pixelConfidence, iou);
 
         #region Helper methods
